Go Sprite added
Added the Go sprite and attempted to change animations by adjusting the "vector2i," which is what I saw when hovering over the frame that adjusted which frame the spritesheet was displaying. I undid all the changes I made but for some reason there's now an invisible wall on the right side of the screen. Need to figure out how to change the spritesheet to display the correct frame when the correct buttons are pressed.